+++ projects/cgmesh/CMakeLists.txt
@@ -7,6 +7,7 @@ add_subdirectory(voro)
 target_link_libraries(zeno_cgmesh PRIVATE voro++)
 
 add_subdirectory(meshfix)
+target_include_directories(zeno_cgmesh PRIVATE meshfix/JMeshExt-1.0alpha_src/include /home/bate/Codes/zeno-blender/external/zeno/projects/cgmesh/meshfix/JMeshExt-1.0alpha_src/JMeshLib-1.2/include)
 
 if (UNIX)
     find_package(LIBIGL QUIET)
@@ -21,7 +22,7 @@ if (UNIX)
         message(FATAL_ERROR "please specify LIBIGL_INCLUDE_DIR")
     endif()
     if (NOT LIBIGL_EIGEN_INCLUDE_DIR)
-        message(FATAL_ERROR "please specify LIBIGL_INCLUDE_DIR")
+        message(FATAL_ERROR "please specify LIBIGL_EIGEN_INCLUDE_DIR")
     endif()
     target_include_directories(zeno_cgmesh PRIVATE ${LIBIGL_INCLUDE_DIR} ${LIBIGL_EIGEN_INCLUDE_DIR})
     target_link_libraries(zeno_cgmesh PRIVATE igl::core igl_cgal Eigen3::Eigen gmp mpfr boost_system boost_log)
